Rename prop interface imports
