+++ src/components/Frame/Frame.tsx
@@ -11,12 +11,9 @@ import Backdrop from '../Backdrop';
 import TrapFocus from '../TrapFocus';
 import {dataPolarisTopBar, layer, Duration} from '../shared';
 import {setRootProperty} from '../../utilities/setRootProperty';
-import {
-  ContextualSaveBarProps,
-  FrameContext,
-  frameContextTypes,
-  ToastProps,
-} from './types';
+import {ContextualSaveBarProps} from '../ContextualSaveBar';
+import {ToastProps} from '../Toast';
+import {FrameContext, frameContextTypes} from './types';
 import {ToastManager, Loading, ContextualSaveBar} from './components';
 
 import * as styles from './Frame.scss';
