+++ src/components/Frame/index.ts
@@ -4,11 +4,6 @@ export {Props} from './Frame';
 
 export {DEFAULT_TOAST_DURATION} from './components';
 
-export {
-  ContextualSaveBarProps,
-  FrameContext,
-  frameContextTypes,
-  ToastProps,
-} from './types';
+export {FrameContext, frameContextTypes} from './types';
 
 export default Frame;
