+++ src/components/Toast/index.ts
@@ -1,3 +1,4 @@
 import Toast from './Toast';
 
+export {ToastProps} from './Toast';
 export default Toast;
